Add contracts
